Add crop node
In this change crop node is implemented.
The purpose of this node is to crop portion of the first node that
dimension-wise corresponds to the second node at the given or
calculated offset.
Offsets may be given manually or calculated based on affine transform
between two inputs along the path that connects them.